Decouple bits by introducing BuilderRegistry
