+++ lib/builder.js
@@ -6,32 +6,6 @@ import path from 'path'
 import LogParser from './parsers/log-parser'
 import MagicParser from './parsers/magic-parser'
 
-let builders = []
-
-function getAllBuilders () {
-  if (builders.length === 0) {
-    const moduleDir = path.join(__dirname, 'builders')
-    const entries = fs.readdirSync(moduleDir)
-    builders = entries.map((entry) => require(path.join(moduleDir, entry)))
-  }
-
-  return builders
-}
-
-function resolveAmbigiousBuilders (builders) {
-  const names = builders.map((builder) => builder.name)
-  const indexOfLatexmk = names.indexOf('LatexmkBuilder')
-  const indexOfTexify = names.indexOf('TexifyBuilder')
-  if (builders.length === 2 && indexOfLatexmk >= 0 && indexOfTexify >= 0) {
-    switch (atom.config.get('latex.builder')) {
-      case 'latexmk': return builders[indexOfLatexmk]
-      case 'texify': return builders[indexOfTexify]
-    }
-  }
-
-  throw Error('Unable to resolve ambigous builder registration')
-}
-
 export default class Builder {
   constructor () {
     this.envPathKey = this.getEnvironmentPathKey(process.platform)
@@ -40,17 +14,6 @@ export default class Builder {
   static canProcess (/* filePath */) {}
   run (/* filePath */) {}
   constructArgs (/* filePath */) {}
-
-  static getBuilder (filePath) {
-    const builders = getAllBuilders()
-    const candidates = builders.filter((builder) => builder.canProcess(filePath))
-    switch (candidates.length) {
-      case 0: return null
-      case 1: return candidates[0]
-    }
-
-    return resolveAmbigiousBuilders(candidates)
-  }
 
   parseLogFile (texFilePath) {
     const logFilePath = this.resolveLogFilePath(texFilePath)
