+++ lib/composer.js
@@ -3,7 +3,7 @@
 import _ from 'lodash'
 import fs from 'fs-plus'
 import path from 'path'
-import {heredoc} from './werkzeug'
+import BuilderRegistry from './builder-registry'
 
 export default class Composer {
   destroy () {
@@ -20,9 +20,9 @@ export default class Composer {
       return Promise.reject(false)
     }
 
-    if (!this.isTexFile(filePath)) {
-      latex.log.warning(heredoc(`File does not seem to be a TeX file
-        unsupported extension '${path.extname(filePath)}'.`))
+    const builder = this.getBuilder(filePath)
+    if (builder == null) {
+      latex.log.warning(`No registered LaTeX builder can process ${filePath}.`)
       return Promise.reject(false)
     }
 
@@ -30,7 +30,6 @@ export default class Composer {
       editor.save() // TODO: Make this configurable?
     }
 
-    const builder = latex.getBuilder()
     const rootFilePath = this.resolveRootFilePath(filePath)
 
     this.destroyErrorMarkers()
@@ -254,6 +253,11 @@ export default class Composer {
   isTexFile (filePath) {
     // TODO: Improve will suffice for the time being.
     return !filePath || filePath.search(/\.(tex|lhs)$/) > 0
+  }
+
+  getBuilder (filePath) {
+    const BuilderImpl = BuilderRegistry.getBuilder(filePath)
+    return (BuilderImpl != null) ? new BuilderImpl() : null
   }
 
   getEditorDetails () {
