+++ lib/latex.js
@@ -24,30 +24,17 @@ export default class Latex {
   constructor () {
     this.createLogProxy()
 
-    defineDefaultProperty(this, 'builder')
     defineDefaultProperty(this, 'logger')
     defineDefaultProperty(this, 'opener')
 
     this.observeOpenerConfig()
-    this.observeBuilderConfig()
   }
 
-  getBuilder () { return this.builder }
   getLogger () { return this.logger }
   getOpener () { return this.opener }
 
   setLogger (logger) {
     this.logger = logger
-  }
-
-  getDefaultBuilder () {
-    let BuilderClass = null
-    if (this.useLatexmk()) {
-      BuilderClass = require('./builders/latexmk')
-    } else {
-      BuilderClass = require('./builders/texify')
-    }
-    return new BuilderClass()
   }
 
   getDefaultLogger () {
@@ -90,11 +77,6 @@ export default class Latex {
     atom.config.onDidChange('latex.skimPath', callback)
     atom.config.onDidChange('latex.sumatraPath', callback)
     atom.config.onDidChange('latex.okularPath', callback)
-  }
-
-  observeBuilderConfig () {
-    const callback = () => { this['__builder'] = this.getDefaultBuilder() }
-    atom.config.onDidChange('latex.builder', callback)
   }
 
   resolveOpenerImplementation (platform) {
@@ -156,9 +138,5 @@ export default class Latex {
 
   viewerExecutableExists () {
     return fs.existsSync(atom.config.get('latex.viewerPath'))
-  }
-
-  useLatexmk () {
-    return atom.config.get('latex.builder') === 'latexmk'
   }
 }
